refactor: move additional_loggers and logging setup to VcsEnvironment
VcsEnvironment now stores additional_loggers and provides
configure_logging() which sets up all loggers at the correct level.
GlobalOverrides.__enter__ delegates to vcs_env.configure_logging().

- Add additional_loggers field to VcsEnvironment
- Add configure_logging() method to VcsEnvironment
- Remove additional_loggers from GlobalOverrides __slots__/__init__
- Add backward-compat property delegating to vcs_env
- from_env() attaches loggers via dataclasses.replace
- Simplify _log._get_all_scm_loggers: no longer reads _active_overrides
